Detect_Homograph_Parking/requirements.txt
onnxruntime>=1.23.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.5.0
ultralytics>=8.0.0
onnx>=1.14.0
onnxsim>=0.4.0
Pillow>=9.0.0
ipywidgets>=7.7.0
ipympl>=0.9.0
